Fix profile genre normalization by applying in JavaScript
- Extract genre mappings and normalization functions to shared utils/genres.js
- Update audiobooks.js to import from shared module (removes duplication)
- Update profile.js stats endpoint to normalize genres in JavaScript
  instead of querying non-existent normalized_genre column
- Properly aggregate listen time by normalized genre categories
